+++ services/dynamic-scheduler/src/simcore_service_dynamic_scheduler/services/director_v2/_thin_client.py
@@ -1,8 +1,8 @@
 import datetime
 from typing import cast
 
+from common_library.exclude import as_dict_exclude_none
 from common_library.json_serialization import json_dumps
-from common_library.unset import as_dict_exclude_none
 from fastapi import FastAPI, status
 from httpx import Response, Timeout
 from models_library.api_schemas_dynamic_scheduler.dynamic_services import (
