+++ services/web/server/src/simcore_service_webserver/folders/_folders_db.py
@@ -10,7 +10,7 @@ from typing import Final, cast
 
 import sqlalchemy as sa
 from aiohttp import web
-from common_library.unset import UnSet, as_dict_exclude_unset
+from common_library.exclude import UnSet, as_dict_exclude_unset
 from models_library.folders import (
     FolderDB,
     FolderID,
