+++ services/web/server/src/simcore_service_webserver/projects/_folders_db.py
@@ -8,7 +8,7 @@ import logging
 from datetime import datetime
 
 from aiohttp import web
-from common_library.unset import UnSet, as_dict_exclude_unset
+from common_library.exclude import UnSet, as_dict_exclude_unset
 from models_library.folders import FolderID
 from models_library.projects import ProjectID
 from models_library.users import UserID
